MainITest refactoring
Moved common code to ExecutableITestBase
Split test cases in separate classes, with meaninful names
